+++ Classes/ASIDataDecompressor.m
@@ -65,7 +65,7 @@
 	return nil;
 }
 
-- (NSData *)uncompressBytes:(Bytef *)bytes length:(NSInteger)length error:(NSError **)err
+- (NSData *)uncompressBytes:(Bytef *)bytes length:(NSUInteger)length error:(NSError **)err
 {
 	if (length == 0) return nil;
 	
@@ -75,7 +75,7 @@
 	int status;
 	
 	zStream.next_in = bytes;
-	zStream.avail_in = length;
+	zStream.avail_in = (unsigned int)length;
 	zStream.avail_out = 0;
 	NSError *theError = nil;
 	
@@ -149,7 +149,7 @@
 	
 	UInt8 inputData[DATA_CHUNK_SIZE];
 	NSData *outputData;
-	int readLength;
+	NSInteger readLength;
 	NSError *theError = nil;
 	
 	ASIDataDecompressor *decompressor = [ASIDataDecompressor decompressor];
